requirements.txt
# Core LangGraph / LangChain
langgraph>=1.0.7
langgraph-supervisor>=0.0.31
langgraph-checkpoint-postgres>=3.0.4
langgraph-checkpoint-sqlite>=2.0.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langsmith>=0.3.0

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0

# Integrations
slack-bolt>=1.20.0
slack-sdk>=3.30.0
atlassian-python-api>=3.41.0
notion-client>=2.2.0
PyGithub>=2.6.0
tavily-python>=0.5.0
praw>=7.8.0
beautifulsoup4>=4.12.0
httpx>=0.27.0

# Scheduler
apscheduler>=3.10.0

# Database
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
sqlalchemy>=2.0.0

# Observability
opik>=1.10.0

# Utilities
pydantic>=2.9.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
structlog>=24.0.0
tenacity>=9.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
respx>=0.22.0
